# LabRequirement3/static/swagger.yaml
openapi: 3.0.0
info:
  title: Todo API
  version: 1.0.0
  description: A simple API to manage Todos
servers:
  - url: http://127.0.0.1:5000
paths:
  /todos/:
    get:
      summary: Get all todos
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    task:
                      type: string
                    completed_task:
                      type: boolean
    post:
      summary: Add a new todo
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                task:
                  type: string
                completed_task:
                  type: boolean
      responses:
        '201':
          description: Todo created
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    task:
                      type: string
                    completed_task:
                      type: boolean
  /todos/{todo_id}:
    put:
      summary: Update a todo by ID
      parameters:
        - in: path
          name: todo_id
          schema:
            type: integer
          required: true
          description: Numeric ID of the todo to update
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                task:
                  type: string
                completed_task:
                  type: boolean
      responses:
        '200':
          description: Todo updated
    delete:
      summary: Delete a todo by ID
      parameters:
        - in: path
          name: todo_id
          schema:
            type: integer
          required: true
          description: Numeric ID of the todo to delete
      responses:
        '204':
          description: Todo deleted successfully